+++ lib/src/widgets/radio.dart
@@ -13,21 +13,22 @@ class ShokoUIRadio extends StatelessWidget {
   final Color? inactiveColor;
   final Color? activeColor;
   final Color? markColor;
-  final ShokoUIRadii radius;
-  final ShokoUIShadow? shadow;
+  final ShokoUIRadii? radius;
 
   final ShokoUIRadioChildPosition childPosition;
+  final MainAxisAlignment mainAxisAlignment;
 
   const ShokoUIRadio({super.key,
     required this.value, this.onChange,
     this.child,
-    this.size = 32, this.inactiveColor, this.markColor, this.activeColor, this.shadow, this.radius = ShokoUIRadii.medium,
-    this.childPosition = ShokoUIRadioChildPosition.right
+    this.size = 32, this.inactiveColor, this.markColor, this.activeColor, this.radius = ShokoUIRadii.medium,
+    this.childPosition = ShokoUIRadioChildPosition.right, this.mainAxisAlignment = MainAxisAlignment.start
   });
 
   @override
   Widget build(BuildContext context) {
     return Row(
+      mainAxisAlignment: mainAxisAlignment,
       mainAxisSize: MainAxisSize.min,
       children: [
         if (childPosition == ShokoUIRadioChildPosition.left) ... [
@@ -45,14 +46,14 @@ class ShokoUIRadio extends StatelessWidget {
             height: size,
             width: size,
             decoration: BoxDecoration(
-              color: value ? (activeColor ?? Colors.red) : null,
-              borderRadius: radius.get(),
+              color: value ? ((activeColor ?? context.shokoTheme?.radioTheme?.activeColor) ?? Colors.red) : null,
+              borderRadius: (radius?.get() ?? context.shokoTheme?.buttonTheme?.radius?.get()),
               border: !value ? Border.all(
-                color: inactiveColor ?? (activeColor ?? Colors.red),
+                color: (inactiveColor ?? context.shokoTheme?.radioTheme?.inactiveColor) ?? ((activeColor ?? context.shokoTheme?.radioTheme?.activeColor) ?? Colors.red),
                 width: 2
               ) : null
             ),
-            child: value ? Icon(Icons.check_rounded, size: size - 4, color: markColor ?? Colors.white) : null,
+            child: value ? Icon(Icons.check_rounded, size: size - 4, color: (markColor ?? context.shokoTheme?.radioTheme?.markColor) ?? Colors.white) : null,
           )
         ),
         if (childPosition == ShokoUIRadioChildPosition.right) ... [