+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+## 0.0.2
+
+Addendum ShokoUIThemeWidget for global theme for other a shoko widget
+
+Supported theme: all widgets
+
+Modefied widgets:
+    radio: remove shadow
+
 ## 0.0.1
 
-* TODO: Describe initial release.
+Supported widgets: button, radio, switch, textfield
+Supported styles: radius, shadow